+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Added
 
 - Added skeleton loading state for listings details page
+- Added Loader component when loading state for favorites & reservations pages
 
 # [0.4.2]